添加模拟交易

+++ internal/domain/portfolio/snapshot.go
@@ -2,7 +2,7 @@ package portfolio
 
 type PositionSnapshot struct {
 	Symbol    string
-	Quantity  int64
+	Quantity  float64
 	AvgPrice  float64
 	LastPrice float64
 }
